+alist_to_pitch,+duration_to_alist
Ignore-this: c9ae3197424836ec97cb5a8c86338945

darcs-hash:60a0b299a12015437ba108baa1a5be8f94a86919

+++ Sound/SC3/Lang/Collection.hs
@@ -594,7 +594,7 @@ t2_window n x =
 -- > t2_adjacent [1..6] == [(1,2),(3,4),(5,6)]
 -- > t2_adjacent [1..5] == [(1,2),(3,4)]
 t2_adjacent :: [t] -> [(t,t)]
-t2_adjacent = t2_window 2
+t2_adjacent = t2_window (2::Int)
 
 -- | List of 2-tuples of overlapping elements.
 --
